extract table columns and mappers
