+++ apps/earn-protocol/components/organisms/RebalancingActivity/columns.ts
@@ -1,0 +1,27 @@
+export const rebalancingActivityColumns = [
+  {
+    title: 'Purpose',
+    key: 'purpose',
+    sortable: false,
+  },
+  {
+    title: 'Action',
+    key: 'action',
+    sortable: false,
+  },
+  {
+    title: 'Amount',
+    key: 'amount',
+    sortable: false,
+  },
+  {
+    title: 'Timestamp',
+    key: 'timestamp',
+    sortable: false,
+  },
+  {
+    title: 'Provider',
+    key: 'provider',
+    sortable: false,
+  },
+]
