+++ apps/earn-protocol/components/organisms/UserActivity/columns.ts
@@ -1,0 +1,27 @@
+export const userActivityColumns = [
+  {
+    title: 'Position Balance',
+    key: 'balance',
+    sortable: false,
+  },
+  {
+    title: 'Amount',
+    key: 'amount',
+    sortable: false,
+  },
+  {
+    title: '# of Deposits',
+    key: 'numberOfDeposits',
+    sortable: false,
+  },
+  {
+    title: 'Time',
+    key: 'time',
+    sortable: false,
+  },
+  {
+    title: 'Earning streak',
+    key: 'earningStreak',
+    sortable: false,
+  },
+]
